Removed `view_count` column from image table
